# Change pendingAuthorizations indexes for better performance.